fix lint warnings in settings panel
