+++ src/components/settings/SettingsSearchBar.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faSearch } from '@/lib/icons';
+
+interface SettingsSearchBarProps {
+  searchTerm: string;
+  onSearchChange: (term: string) => void;
+}
+
+const SettingsSearchBar: React.FC<SettingsSearchBarProps> = ({ searchTerm, onSearchChange }) => {
+  return (
+    <div className="relative mb-6">
+      <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+        <FontAwesomeIcon icon={faSearch} className="text-gray-400" />
+      </div>
+      <input
+        type="text"
+        placeholder="Search settings..."
+        value={searchTerm}
+        onChange={(e) => onSearchChange(e.target.value)}
+        className="w-full pl-10 pr-4 py-3 bg-gray-800 border border-gray-700 rounded-lg text-gray-200 placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all duration-200"
+        aria-label="Search settings"
+      />
+    </div>
+  );
+};
+
+export default SettingsSearchBar;
+
